+++ test/integration/targets/truffle/migrations/1_initial_migration.ts
@@ -3,7 +3,7 @@ const DumbContract = artifacts.require("DumbContract");
 
 module.exports = function(deployer) {
   deployer.deploy(Migrations);
-  deployer.deploy(DumbContract);
+  deployer.deploy(DumbContract, 0);
 } as Truffle.Migration;
 
 export {};