+++ test/integration/targets/truffle/test/DumbContract.ts
@@ -4,16 +4,19 @@ const DumbContract = artifacts.require("DumbContract");
 
 contract("DumbContract", ([deployer]) => {
   it("should work", async () => {
-    const contract = await DumbContract.new({ from: deployer });
+    const contract = await DumbContract.new(0, { from: deployer });
 
+    expect(DumbContract.contractName).to.be.a("string");
+
+    expect(contract.address).to.be.a("string");
     expect((await contract.counterWithOffset(2)).toNumber()).to.be.eq(2);
     expect((await contract.returnAll()).map(x => x.toNumber())).to.be.deep.eq([0, 5]);
   });
 
   it("should allow to pass unsigned values in multiple ways", async () => {
-    const contract = await DumbContract.new({ from: deployer });
+    const contract = await DumbContract.new(0, { from: deployer });
 
-    await contract.countup(2);
+    expect((await contract.countup(2)).tx).to.not.be.undefined;
     expect((await contract.counter()).toNumber()).to.be.eq(2);
     await contract.countup(new BigNumber(2));
     expect((await contract.counter()).toNumber()).to.be.eq(4);
@@ -22,7 +25,7 @@ contract("DumbContract", ([deployer]) => {
   });
 
   it("should allow to pass signed values in multiple ways", async () => {
-    const contract = await DumbContract.new({ from: deployer });
+    const contract = await DumbContract.new(0, { from: deployer });
 
     expect((await contract.returnSigned(2)).toNumber()).to.be.eq(2);
     expect((await contract.returnSigned(new BigNumber(2))).toNumber()).to.be.eq(2);
@@ -30,7 +33,7 @@ contract("DumbContract", ([deployer]) => {
   });
 
   it("should allow to pass address values in multiple ways", async () => {
-    const contract = await DumbContract.new({ from: deployer });
+    const contract = await DumbContract.new(0, { from: deployer });
 
     expect(await contract.testAddress("0x123")).to.be.eq(
       "0x0000000000000000000000000000000000000123",
@@ -41,7 +44,7 @@ contract("DumbContract", ([deployer]) => {
   });
 
   it("should allow to pass bytes values in multiple ways", async () => {
-    const contract = await DumbContract.new({ from: deployer });
+    const contract = await DumbContract.new(0, { from: deployer });
 
     expect(await contract.callWithBytes("0x123")).to.be.deep.eq(
       "0x1230000000000000000000000000000000000000000000000000000000000000",
@@ -52,7 +55,7 @@ contract("DumbContract", ([deployer]) => {
   });
 
   it("should allow to pass numeric arrays values in multiple ways", async () => {
-    const contract = await DumbContract.new({ from: deployer });
+    const contract = await DumbContract.new(0, { from: deployer });
 
     expect(
       (await contract.callWithArray(["1", 2, new BigNumber(3)])).map(x => x.toNumber()),
@@ -60,7 +63,7 @@ contract("DumbContract", ([deployer]) => {
   });
 
   it("should allow to pass strings ", async () => {
-    const contract = await DumbContract.new({ from: deployer });
+    const contract = await DumbContract.new(0, { from: deployer });
 
     expect(await contract.testString("abc")).to.be.deep.eq("abc");
   });